+++ codegen/src/lib.rs
@@ -3,33 +3,27 @@
 #![recursion_limit = "1024"]
 
 pub use crate::{
-    abi::{ToLSBytes, Type},
     asm::Assembler,
     code::Code,
     codegen::CodeGen,
     constructor::Constructor,
     control::{ControlStack, ControlStackFrame, ControlStackFrameType},
-    data::DataSet,
     dispatcher::Dispatcher,
-    func::{Function, Functions},
     jump::JumpTable,
     local::{LocalSlot, Locals},
     masm::MacroAssembler,
     result::{Error, Result},
-    wasm::{Exports, HostFunc, Imports},
+    wasm::{Data, Exports, Function, Functions, HostFunc, Imports, ToLSBytes, Type},
 };
 use smallvec::SmallVec;
 
-pub mod abi;
 mod asm;
 mod backtrace;
 mod code;
 mod codegen;
 mod constructor;
 mod control;
-mod data;
 mod dispatcher;
-mod func;
 mod jump;
 mod local;
 mod masm;
